Refactor Spiffy order bumps
This preps order bumps to be included within Spiffy Core. We're now using the same methodology as our translation snippet with defaults being set. This also adds the ability to have bumps default to ENABLED on initial page load. We need to revisit this on locked down order forms to ensure we're not changing the order amount.